chore: rename to ssr-handlers
