+++ packages/core/index.ts
@@ -106,4 +106,4 @@ export * from './useWindowScroll'
 export * from './useWindowSize'
 export * from './types'
 export * from '@vueuse/shared'
-export * from './ssr-context'
+export * from './ssr-handlers'
